feat: add encoded histogram export (#84)
A work-around solution to #82, adding a feature to export the underlying
histograms using their codable conformance. The preferred longer term
solution requires additional features/capabilities on package-histogram
to support cross-platform encoding and decoding of the raw histogram
data.

This encoded histogram *does not* include the units of measurement
selected for the various benchmarks but is in raw form (which is `nanoseconds` for time measurements and just `count` for countable measurements).