feat(minor): Make CI fail with regression for crashes or other run failures (#211)
After the changes in https://github.com/ordo-one/package-benchmark/pull/166 we no longer fail the run of the rest of the benchmarks if a single one fails, which is reasonable during normal iterative work. This is not ok for CI baseline checks though, so this PR restores failures when `baseline check` is performed.