fix: memory unsafe file read on Linux
When reading from the file descriptor into the temporary buffer, it was
only by happenstance that the buffer would contain a terminating nul
byte.

The buffer is uninitialized and thus may contain any data.

And the `/proc` files that are read into the buffer are **not** nul
terminated.

If the temporary buffer happened to be zero-initialized, then this would
accidentally work correctly.

If the temporary buffer happend to have a nul byte somewhere after the
read data and in-bounds, then the end of the resulting String would be
corrupted with arbitrary crud. This crud would ostensibly be after a
terminating newline and thus parsers might have ignored it?

If the temporary buffer happened to not have any nul bytes after the
read data, then other Bad Things could happen, including silent success
or a crash due to segmentation fault or various checks in String being
violated.

The repair is to append a nul byte after the read data (ensuring that
the nul byte is not placed after the end of the buffer).

Signed-off-by: Peter Grayson <pete@jpgrayson.net>