feat(minor): Add Benchmark parameters (#274)
## Description

Sometimes it is convenient to run the same benchmark multiple times
under slightly different scenarios. For example, benchmarking `func
foo(_ n: Int)` for various parameterizations of `n`. This adds support
for specifying tags for a benchmark but makes no functional
changes beyond the name/identifier of a benchmark now including a
parameter list when one was provided.

Co-authored-by: Jairon Terrero <jairon@passivelogic.com>